Create README.md
Simple program that stores encrypted passwords

I made this for myself, I was tired of having all my passwords on a paper for keeping them secure. A paper will always be the most secure way to store passwords (you can't hack a paper :D), but it's not as handy as a text field in your PC that can be copy/pasted. So, this program isn't better than the glorious paper in terms of security, but it will keep your passwords encrypted, protected by a master password defined by you. It's better than having them in a notepad file like I used to do until I got hacked :/